Move response timeout to http package
