+++ aws/transport/http/timeout_read_closer.go
@@ -1,4 +1,4 @@
-package aws
+package http
 
 import (
 	"context"
